Include libraries and modify prototype

+++ 0x0E-linear_skip/search.h
@@ -1,6 +1,9 @@
 #ifndef SEARCH_H
 #define SEARCH_H
 
+#include <stdio.h>
+#include <stddef.h>
+
 /**
  * struct skiplist_s - Singly linked list with an express lane
  *
@@ -20,7 +23,7 @@ typedef struct skiplist_s
     struct skiplist_s *express;
 } skiplist_t;
 
-int linked_list_len(skiplist_t *list);
+size_t linked_list_len(skiplist_t *list);
 skiplist_t *linear_skip(skiplist_t *list, int value);
 
 #endif
